Implemented Guardians functionality
Encryption and mounted storages scanning mechanisms are introduced.
A Guardian can now be added for real. A pass-file with name, key, and
decrypted message is written to a storage, whereas database holds name
and encrypted message.
The app now monitors mounted storages in real time and recognizes if
mounted storages are Guardians. Mounted storages are monitored in a
parallel thread.