Implemented items and system protection
A major update: most app features are implemented and there is not so
much to be done. One can already pair an item to a guardian or create a
system key. While the key is mounted, one can enable\disable protection
so that the port can be used for other needs. The source code was
remastered so that the app is split into independent modules now (each
folder stands for a separate module). The interaction between modules is
implemented inside a MainController class. GUI is remastered too (a
little bit).